customer-service-assistant/requirements.txt
openai==1.3.5
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
python-multipart==0.0.6
numpy==1.26.2
